requirements.txt
fastmcp
azure-devops
python-dotenv

# Testing
pytest
pytest-asyncio
pytest-mock